pyproject.toml
[tool.poetry]
name = "doc-generator"
version = "0.1.0"
description = "A Python-based tool for generating .NET XML documentation in C# code."
authors = ["klebowski <klebowskimaciej@gmail.com>"]
license = "MIT"
readme = "README.md"

packages = [
  { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.0.0"
transformers = "^4.26.0"
python-dotenv = "^1.0.0"
google-generativeai = "^0.1.0"

[tool.poetry.scripts]
doc-generator = "src.cli:main"

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"